Update main.py
Switched from using lists to dictionaries for creating the deck. This required an small changes to how we would get a card from the deck. 
Implemented 3 values within each card: display name, suit, value. Display name is the name displayed in the game (ace, 2, 3,... king). Suit is the 4 suits (e.g. Diamond). Value is the hidden value of the card (royal names are valued as 10).